Actualizadas practicas y tareas de Web Development. Tambien agregado prototipo de Tarea7 en Distribuidos y Practica6 en Admin. Redes.

+++ java/SistemasDistribuidos/Tareas/Tarea7-Comercio/Servicio/servicio_json/Articulo.java
@@ -5,7 +5,7 @@ import java.sql.Timestamp;
 public class Articulo{
   String nombre;
   String descripcion;
-  float precio;
+  int precio;
   int cantidad;
   byte[] foto;
 }
